Trait mp_lab_ii::lr_3::base::Tree

pub trait Tree<T: Ord + Copy + Debug, TN: TreeNode<T>> {

    // Required methods
    fn get_root(&self) -> &Option<Rc<RefCell<TN>>>;
    fn insert(&mut self, data: T);
    fn delete(&mut self, data: T);
    fn print_tree(&self);

    // Provided methods
    fn get_height(&self) -> u32 { ... }
    fn get_min(&self) -> Option<T> { ... }
    fn get_max(&self) -> Option<T> { ... }
    fn count_leaves(&self) -> u32 { ... }
    fn count_nodes(&self) -> u32 { ... }
    fn traverse_inorder(&self) { ... }
    fn traverse_preorder(&self) { ... }
    fn traverse_postorder(&self) { ... }
    fn contain(&self, value: T) -> bool { ... }
    fn is_empty(&self) -> bool { ... }

}

Required Methods

fn get_root(&self) -> &Option<Rc<RefCell<TN>>>

fn insert(&mut self, data: T)

fn delete(&mut self, data: T)

fn print_tree(&self)

Provided Methods

fn get_height(&self) -> u32

The height of the tree.

This function can be used in RBTree, AVLTree and BSTree.

Returns to the height.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.get_height(), 0);

tree.insert(1);
tree.insert(0);
tree.insert(2);
assert_eq!(tree.get_height(), 2);

fn get_min(&self) -> Option<T>

The minimum element of the tree.

This function can be used in RBTree, AVLTree and BSTree.

Returns to the minimum element.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.get_min(), None);

tree.insert(1);
tree.insert(0);
tree.insert(2);
assert_eq!(tree.get_min(), Some(0));

fn get_max(&self) -> Option<T>

The maximum element of the tree.

This function can be used in RBTree, AVLTree and BSTree.

Returns to the maximum element.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.get_max(), None);

tree.insert(1);
tree.insert(0);
tree.insert(2);
assert_eq!(tree.get_max(), Some(2));

fn count_leaves(&self) -> u32

Returns the number of leaf nodes in the tree.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.count_leaves(), 0);
tree.insert(1);
tree.insert(2);
assert_eq!(tree.count_leaves(), 1);

fn count_nodes(&self) -> u32

Returns the number of elements in the tree.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.count_nodes(), 0);
tree.insert(1);
assert_eq!(tree.count_nodes(), 1);

fn traverse_inorder(&self)

Inorder traverse iterator of tree.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree= RBTree::new();
    let v=vec![1,2,3,4,5,6,7];
       for i in v {
           tree.insert(i);
       }

// Now we have a tree that looks like this:
//                  2
//               1     4
//                   3   6
//                      5 7

// And we should get the following sequence of its elements: 1, 2, 3, 4, 5, 6, 7
tree.traverse_inorder();

fn traverse_preorder(&self)

Preorder traverse iterator of tree.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree= RBTree::new();
    let v=vec![1,2,3,4,5,6,7];
       for i in v {
           tree.insert(i);
       }

// Now we have a tree that looks like this:
//                  2
//               1     4
//                   3   6
//                      5 7

// And we should get the following sequence of its elements: 2, 1, 4, 3, 6, 5, 7
tree.traverse_preorder();

fn traverse_postorder(&self)

Postorder traverse iterator of tree.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree= RBTree::new();
    let v=vec![1,2,3,4,5,6,7];
       for i in v {
           tree.insert(i);
       }

// Now we have a tree that looks like this:
//                  2
//               1     4
//                   3   6
//                      5 7

// And we should get the following sequence of its elements: 1, 3, 5, 7, 6, 4, 2
tree.traverse_postorder();

fn contain(&self, value: T) -> bool

Checks whether the tree contains an element with the specified value.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert_eq!(tree.contain(1), false);

tree.insert(1);
tree.insert(0);
tree.insert(2);
tree.insert(1);

assert!(tree.contain(2));
assert!(tree.contain(1));
assert!(!tree.contain(999));

fn is_empty(&self) -> bool

Сhecking if the tree is empty.

This function can be used in RBTree, AVLTree and BSTree.

Examples

use trees::rbtree::RBTree;
use crate::trees::base::Tree;

let mut tree = RBTree::new();
assert!(tree.is_empty());

tree.insert(1);
assert!(!tree.is_empty());